requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.7.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.2
pymongo==4.13.0
python-dotenv==1.1.0
typing_extensions==4.14.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
